meant to work on orbit controls
instead reworked the backdrop to be tiles of random colors, mostly
purple. also reworked light angle to better see ship shadow. this work
is incremental, but should be a good step toward hills etc, as well
as work on chunking content to render so the map can grow a lot.