the laser now connects the two points like one might expect
currently the vertices for the laser are computed on every update,
so way too much. also the laser vertices are not quite right and
there is some crossing. and there are no laser effects over time
yet. and there is no laser sound. and nothing happens when a
laser hits the enemy. and i did not move this version of the
laser logic over to the new laser.rs file yet. but you can
see a laser kind of fire! and the transforms all work.
so that's nice.