lasers now have a max range and explode the bot
when the bot is inside the max range, the bot explodes.
i'm still seeing some lag, likely from a few things but mostly
running too much logic during update. the laser box is perpetually
being recomputed, even when it is not being fired. but that was
an issue before this change. a lot of my logic could run ~10
times a second rather than at like 60+ times a second.
the sound is lagging too.