for the sake of performance, i have now made the plaidsea boring temporarily.
but it will be exciting again soon*. notably this does not solve another
performance problem i have wherein i get fps drops to 45 or so on a new
laptop even when fully zoomed-in. that will require further work.
but it does resolve the fps drops to 10 fps when zoomed out.