the bot now fires a laser at the player every few seconds
the laser doesn't yet destroy the player. reworked is_playing
for the laser sound as a dictionary to handle each player's instance
of it playing. made pilot implement debug. time_delta is now assigned
once at the beginning of the more complex functions. although supposedly
multiple calls to get_delta() or similar were fine, it is just more
clear this way that the value is unchanged.